remove dashboard related code
