--- .github/workflows/dashboard.yml
@@ -1,26 +1,0 @@
-name: Dashboard
-
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [master]
-  workflow_dispatch: {}
-
-defaults:
-  run:
-    working-directory: ./references/dashboard
-jobs:
-  build:
-    runs-on: ubuntu-20.04
-    strategy:
-      matrix:
-        node-version: [12.x]
-
-    steps:
-    - uses: actions/checkout@v2
-    - name: Use Node.js ${{ matrix.node-version }}
-      uses: actions/setup-node@v1
-      with:
-        node-version: ${{ matrix.node-version }}
-    - run: npm install
